+++ CHANGELOG.md
@@ -4,8 +4,23 @@ It is organized by version and release date followed by a list of Enhancements, 
 <br /><br />
 
 
-## Version 0.0.4 (Latest) 
-**Releasd:** May 4, 2023<br />
+## Version 0.0.5 (Latest) 
+**Released:** October 26, 2023<br />
+**Tag:** v0.0.5
+
+### Enhancements
+
+- Removed two methods from the "normalization" module:
+  - `remove_leading_trailing_spaces`: This method was used to remove leading and trailing spaces in the input text. (Method removed in this release)
+  - `replace_multiple_spaces`: This method was used to convert consecutive double spaces into single spaces in the input text. (Method removed in this release)
+
+- Added a new method to the "normalization" module:
+  - `remove_whitespace(text)`: This new method efficiently removes all excess spaces in the input text. It replaces consecutive sequences of spaces with a single space and removes any leading or trailing spaces, ensuring a cleaner and more consistent text output.
+
+<br /><br />
+
+## Version 0.0.4 
+**Released:** May 4, 2023<br />
 **Tag:** v0.0.4
 
 ### Enhancements
@@ -20,7 +35,7 @@ It is organized by version and release date followed by a list of Enhancements, 
 <br /><br />
 
 ## Version 0.0.3
-**Releasd:** May 2, 2023<br />
+**Released:** May 2, 2023<br />
 **Tag:** v0.0.3
 
 ### Bug Fix
@@ -30,7 +45,7 @@ It is organized by version and release date followed by a list of Enhancements, 
 <br /><br />
 
 ## Version 0.0.2 
-**Releasd:** May 1, 2023<br />
+**Released:** May 1, 2023<br />
 **Tag:** v0.0.2
 
 ### General Changes
@@ -43,8 +58,9 @@ It is organized by version and release date followed by a list of Enhancements, 
 - Fixed an unidiomatic-typecheck (C0123) from type() to isinstance(). The idiomatic way to perform an explicit typecheck in Python is to use isinstance(x, y) rather than type(x) == Y.
 
 <br /><br />
+
 ## Version 0.0.1 (Initial Release)
-**Releasd:** April 28, 2023<br />
+**Released:** April 28, 2023<br />
 **Tag:** v0.0.1
 
 This is the initial release